test(storybook): colocate storybook test stories & playwright tests
