+++ packages/storybook/playwright.config.ts
@@ -2,7 +2,7 @@ import { defineConfig, devices } from '@playwright/test'
 
 export default defineConfig({
   // Look for test files in the "tests" directory, relative to this configuration file.
-  testDir: 'tests',
+  testDir: 'src/tests',
 
   // Run all tests in parallel.
   fullyParallel: true,
