+++ packages/storybook/src/tests/focus-on-select/index.test.ts
@@ -1,8 +1,8 @@
 import { test, expect } from '@playwright/test'
-import { examplePath } from './test-helpers'
+import { testPath } from '../utils'
 
 test('slide navigation', async ({ page }) => {
-  await page.goto(examplePath('focus-on-select-test'))
+  await page.goto(testPath('focus-on-select'))
   const dots = await page.locator('.v-slick-dots > *')
   await expect(dots.nth(0)).toHaveClass('active')
   await page.waitForTimeout(1000)
