+++ packages/storybook/src/tests/one-slide-finite/index.test.ts
@@ -1,26 +1,25 @@
 import { test, expect } from '@playwright/test'
-import { examplePath } from './test-helpers'
+import { testPath } from '../utils'
 
 test('slide navigation', async ({ page }) => {
-  await page.goto(examplePath('one-slide-infinite-test'))
+  await page.goto(testPath('one-slide-finite'))
   const dots = await page.locator('.v-slick-dots > *')
   await expect(dots.nth(0)).toHaveClass('active')
   await page.waitForTimeout(1000)
   await expect(dots.nth(0)).not.toHaveClass('active')
-  await expect(dots.nth(9)).toHaveClass('active')
+  await expect(dots.nth(1)).toHaveClass('active')
   await page.waitForTimeout(1000)
   await expect(dots.nth(0)).not.toHaveClass('active')
-  await expect(dots.nth(9)).not.toHaveClass('active')
-  await expect(dots.nth(8)).toHaveClass('active')
-  await page.waitForTimeout(1000)
-  await expect(dots.nth(0)).not.toHaveClass('active')
-  await expect(dots.nth(9)).toHaveClass('active')
-  await expect(dots.nth(8)).not.toHaveClass('active')
-  await page.waitForTimeout(1000)
-  await expect(dots.nth(0)).toHaveClass('active')
-  await expect(dots.nth(9)).not.toHaveClass('active')
-  await expect(dots.nth(8)).not.toHaveClass('active')
+  await expect(dots.nth(1)).not.toHaveClass('active')
+  await expect(dots.nth(2)).toHaveClass('active')
   await page.waitForTimeout(1000)
   await expect(dots.nth(0)).not.toHaveClass('active')
   await expect(dots.nth(1)).toHaveClass('active')
+  await expect(dots.nth(2)).not.toHaveClass('active')
+  await page.waitForTimeout(1000)
+  await expect(dots.nth(0)).toHaveClass('active')
+  await expect(dots.nth(1)).not.toHaveClass('active')
+  await expect(dots.nth(2)).not.toHaveClass('active')
+  await page.waitForTimeout(1000)
+  await expect(dots.nth(0)).toHaveClass('active')
 })
