--- packages/storybook/tests/test-helpers.ts
@@ -1,5 +1,0 @@
-export const scopeId = 'tests'
-
-export function examplePath(id: string): string {
-  return `/iframe.html?id=${scopeId}-${id.split('-').join('')}--${id}`
-}
